GHub

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: gmarsi <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/23 18:40:19 by gmarsi            #+#    #+#             */
-/*   Updated: 2020/01/28 22:27:54 by gmarsi           ###   ########.fr       */
+/*   Updated: 2020/02/07 19:07:49 by gmarsi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 	while (i < n)
 	{
 		if (str1[i] != str2[i])
-			return ((unsigned int)str1[i] - (unsigned int)str2[i]);
+			return (str1[i] - str2[i]);
 		i++;
 	}
 	return (0);
